query_preprocessing/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.4
distlib==0.3.7
distro==1.9.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.12.4
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
groq==0.4.2
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.40
langchain-google-genai==0.0.9
langchain-groq==0.1.0
langchain-anthropic
langsmith==0.1.3
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
platformdirs==3.10.0
passlib==1.7.4
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pypdf==4.0.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
yarl==1.9.4
